requirements.txt
discord.py==1.7.3
PyMySQL==1.0.2
Flask==2.0.2